# fix: Reflect review
- `ColorComponent`를 분리해야 타입 추론이 정상적으로 되므로(아마 타입을 `PropType` 같은 것으로 주입하여 주는 듯) 위치만 변경.
- `ESLint` 설정 초기화 후 `// eslint-disable-next-line`
