+++ .eslintrc
@@ -31,12 +31,6 @@
     "import/no-unresolved": "off",
     "prettier/prettier": "error",
     "@typescript-eslint/no-unused-vars": "error",
-    "@typescript-eslint/no-empty-interface": [
-      "error",
-      {
-        "allowSingleExtends": true
-      }
-    ],
     "import/no-extraneous-dependencies": ["warn", { "devDependencies": true }]
   },
   "ignorePatterns": ["node_modules"]
