+++ src/themes/types/emotion.d.ts
@@ -1,5 +1,6 @@
 import type { EmotionTheme } from '@themes/emotion-theme';
 
 declare module '@emotion/react' {
+  // eslint-disable-next-line
   export interface Theme extends EmotionTheme {}
 }
